Treeya360: Update Synaptics FW with PID 7817
For the project Grunt-Treeya360,Mutto+AUO,PID:1A21
When the customer accidentally interrupts the fw upgrade of the \
touchscreen,the fw update fails,it will enter the recovery state, \
and the PID will become 7817,the touchscreen will not work.
So as a short-term solution,update fw here to recover from PID 7817 to PID 1A21.

BUG=b:160813586
TEST=1.cros_workon-grunt start chromeos-config-bsp-grunt
     2.emerge-grunt chromeos-config chromeos-bsp-grunt chromeos-config-bsp-grunt
     3.rebuild packages,install image to DUT,check the FW can be updated successfully.

Change-Id: Ib5a7ce6b1c855d93a2dfd79eae6083eae081e5cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2411993
Tested-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>