octopus: revise the term - mic of audio in model.yaml to loc
In model.yaml, there is a term - mic in audio section to indicate
whether there are the one or two mic setting(s) in HiFi.conf. Therefore
the meaning of this variable should present to how many locations of
mics are located to. Only two numbers are possible - 1 is for DUT with
front camera only and 2 is for DUT with front + rear cameras.

BUG=b:168573592
TEST=`emerge-octopus chromeos-config-bsp-octopus chromeos-config`
TEST=`emerge-octopus chromeos-bsp-octopus`

Change-Id: I3180eeea8ffdad153561695591a8f9919c6913d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2426623
Reviewed-by: Henry Sun <henrysun@google.com>
Commit-Queue: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>