baseboard-trogdor: Add udev rule to enable USB wakeup
Set the wakeup flag of the XHCI controller to enable USB wakeup.
On SC7180 this isn't strictly needed for USB wakeup to work (the
dwc3-qcom driver is the relevant part), however the USB hub
driver on trogdor will evaluate the wakeup flag of the controller
(which is the XHCI device) to determine whether wakeup is enabled.

BUG=b:120563698
TEST=USB wakeup works on trogdor with hub driver

Change-Id: I6eaaa8d2215cc0115cfa6651bcf360a9dd313bcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2429397
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>