elm/hana: add kernelnext minimum backlight level
Hana and elm devices on kernelnext were using a min (1/16) brightness
level that was too dim, or on some skus not even turning the backlight
on.

Changing this value also changed the default brightness level.
So tweak that value to maintain the same real brightness.

BUG=b:169685670
TEST=build, deploy to elm. test manually on hana and telesu.

Change-Id: Id217b1ce906930ba5c726553a205408210b6bcf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2455767
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Commit-Queue: Evan Benn <evanbenn@chromium.org>
Tested-by: Evan Benn <evanbenn@chromium.org>