Add docker and its depended packages to satlab.
Adding docker and its friends packages in order to run drone
docker images on SatLab. All these packages are almost standard
install from gentoo and configured same as moblab which as one patch
to add rootdir.patch as by default the directory chosen
if for some config tmp files is in a read only partition,
there is no way to pass a param or set an environment variable
to change this path.  So the code was changed from /etc/docker
to /var/run/docker


Change-Id: I907217d08cbd2380edcd454e1b4ade2e6cfb7865
Bug:b:169438179
TEST=build local satlab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2444115
Commit-Queue: Prasad Vuppalapu <prasadv@chromium.org>
Tested-by: Prasad Vuppalapu <prasadv@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>