afdo: Rename CWP profiles
We want to rename "silvermont"/"airmont" profiles to "atom", and
"broadwell" profiles to "bigcore", for ease of future upgrades.

This also simplifies the use of overlays: we just need to define
where we want to use "bigcore" because "atom" will be the default.

Another CL will come to define more boards to use bigcore, but for
this CL, we just migrate from what we have nowadays.

BUG=chromium:1135245
TEST=Will be tested

Change-Id: Id99db61d39f208ac2a2808b95752eae6b068f617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2466615
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Tiancong Wang <tcwang@google.com>
Commit-Queue: Tiancong Wang <tcwang@google.com>