oak: Add kernel module param to bt driver
Due to a problem with the firmware on the Marvell BT controller, we are
seeing a large number of suspend failures on platforms with MVL8897 and
MVL8997. To get around this, we can power down the controller before
suspend and power it back up after suspend. This change modifies the
kernel module parameter necessary to enable this quirk.

BUG=b:168152910
TEST=Ran suspend_stress_test with and without peers 500 times.

Cq-Depend: chromium:2530520
Change-Id: Ie7c15abeb3a84b5f75d16b787ec4c341b6555e4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2538390
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>