trogdor: Allow verify_fsg to start a bit earlier
It can sometimes take a little bit of time for verify_fsg to run.
Let's allow it to get started earlier instead of waiting until rmtfs
wants to start and then blocking it.

NOTE: this requires a co-dependent change in rmtfs so that it won't
start until we're finished since the old stanza used to block rmtfs.

BUG=b:162465221
TEST=w/ all patches, WiFi starts a bit earlier

Cq-Depend: chromium:2544917
Change-Id: I605e8039eba549cd2360c0408ce4e6991d64ba1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2545290
Reviewed-by: Evan Green <evgreen@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>