trogdor: No need to preload qcom drivers since they're now builtin
In order to avoid double-triggering udev, we've decided to move the
qcom drivers to be builtin to the kernel.  See CL:2545233.  We can now
remove them from here.

NOTE: this was also in slightly the wrong place.  It should have been
in baseboard-trogdor, not overlay-trogdor.

BUG=b:162465221
TEST=w/ all patches, WiFi starts a bit earlier

Cq-Depend: chromium:2545233
Change-Id: I32fd65625e0e74f4410991b9824f9c95953828fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2545289
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>