baseboard-trogdor: Use 'step_wise' governor for GPU thermal zones
Currently the trogdor kernel uses the power allocator (IPA) as default
thermal governor. This governor is only suitable for thermal zones with
cooling devices that provide an energy model, which currently isn't the
case for the GPU on sc7180. Use 'step_wise' as thermal governor for the
GPU zones.

BUG=b:171244475
TEST=on lazor:
  cat /sys/class/thermal/thermal_zone1[34]/policy
    step_wise
    step_wise

Change-Id: I8f59ee0361ec60885a4f9dac346bcc2424730f83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2555779
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>