Add draw* WP script for dogfood and lab devices
There will be an upcoming FW rev. The RO section in this new firmware
won't be compatible with older silicon. Introduce an init script that
detects that older hardware and enables the software write protect.
As long as this change is deployed before that poisoned firmware,
dogfooders will be able to continue to use their devices and get RW
firmware updates. RO will be locked for them.

The lab for the most part won't be affected by this because those
devices will sit back on FW 13590 until newer devices are deployed. So
if this script goes off wrong on the lab devices, it should only affect
older devices that are already being replaced anyway.

BUG=b:169626786
TEST=Run on a Drawcia and Drawlat, R88 and R87.

Change-Id: Id731b486d88636cc2c46d6419072867b11d5b39b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2559434
Tested-by: Evan Green <evgreen@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Auto-Submit: Evan Green <evgreen@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>