chipset-{tgl,jsl}: Update IPU6/IPU6SE libraries to version 2020.11.25
Fix in algo library for ITS denoise issues due to tuning parameters:

"Originally we have the ITS denoise test passed. But with tuning
updates, we find it is sensitive to tuning denoise parameters. The root
cause is improper parameter calculation in algorithm when mapping manual
strength setting for denoise. We have the algo fix in this release,
which ensure stable test result regardless of tuning parameters."

BUG=b:149069716
TEST=emerge-volteer intel-ipu6-libs-bin
TEST=emerge-dedede intel-ipu6se-libs-bin

Change-Id: I81ff142d8f5343d63a76cc948a5ad0a694af3c30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2560020
Tested-by: Tian Shu Qiu <tian.shu.qiu@intel.corp-partner.google.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Commit-Queue: Ren-Pei Zeng <kamesan@chromium.org>