lakitu: Make sure stpcpy is defined in kernels
We have been notified by the CrOS toolchain team that recent changes
will result in our kernels failing to compile. However, COS has moved to
different infrastructure hosted on https://cos.googlesource.com, meaning
that these changes won't block COS releases until our next sync with
ChromeOS's sdk/toolchain.

We are making this change because it is simple and will make lakitu-cq
green again on CrOS. However, we should consider disabling the lakitu-cq
building on CrOS because it is not a priority for the COS team to keep
that builder green anymore. The drawback of that though is we would learn
of breaking changes like this much later.

BUG=b:174685352
TEST=cq
RELEASE_NOTE=None

Change-Id: Ie65cd353e74b0844d96d27722028a0bffecb0b76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2568588
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Vaibhav Rustagi <vaibhavrustagi@google.com>
Commit-Queue: Robert Kolchmeyer <rkolchmeyer@google.com>