Enable WPA3-SAE on latest platforms
These all have the latest kernels, with WiFi chips/drivers that support
it.

NB: WPA3-SAE support is runtime detectable to some extent (and
wpa_supplicant does this), but we still want to phase this in as we see
fit. We may eventually drop the USE flag.

Check if wpa_supplicant thinks your driver supports it with:

  ## Will include the text "SAE" for supported drivers.
  sudo -u wpa -g wpa wpa_cli get_capability key_mgmt

BUG=b:172212319
TEST=wifi.* tests

Change-Id: Id65535d11977ebc7f1999f8f25de8d3324c68ad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2585946
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>