chipset-bdw: add VP8 decoding to autotest caps for BDW
crrev.com/c/2612401 enables VP8 hw decoding on Broadwell. This CL
adds the necessary hw_dec_vp8_* capabilities to the yaml file for
Tast tests to recognize this platform as supporting those. The
supported values in particular are the same as in chipset-bsw,
that is the celeron version of BDW.

BUG=b:172338625
TEST=Verified locally on samus running video.DecodeAccel(|VD|VDPerf|Perf)

Change-Id: Icb2ba4e80f07c8c60e7af620e9e4c5921524bbc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2613394
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>