chromeos-bsp-baseboard-oak: update symlinks for VP8 encoders
The non-upstream kernel was using a different driver name for both
encoder nodes on MT8173, even though they are actually the same driver.
This is not how things will work with upstream, so update the udev rules
to be able to differentiate the encoder instances using their DT node
instead.

We also keep the old rule for backward compatibility, especially with
kernels < 4.19 which won't be updated to use the same name.

BUG=b:149437087
TEST= /dev/video-enc0 and /dev/video-enc1 appear on Hana after using the
same driver name for both encoder nodes.

Change-Id: I459b3c968e59cb48575acc51495409a326a81acc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2673592
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>