arc-codec-*: Limit concurrent ARC++ HW encoder instances to eight
Formerly, we allow Android apps to run as many ARC++ HW encoders
as underlying implementation (e.g. drivers and firmware) can.
However, it is not so reasonable that an app runs a lot of HW
encoders simultaneously. So that we limits the concurrent ARC++
HW encoder instances to eight in Chrome. This configuration file
change are required to pass CTS tests with the limitation.

BUG=b:164990427, b:168422427
TEST=CtsMediaTestCases on elm-kernelnext

Change-Id: Ibd4ae8934fda0a679830baeeec3cf53dd505594b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2569857
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>