jinlon: reduce the LTE modem reboot delay when wedged
Tweak the delay after which we check whether the LTE modem has appeared
on the USB bus and reset it if it is wedged.
Reduce it to 20 seconds, so we expedite the recovery when the modem is
stuck after a warm reboot on Jinlon.

BUG=b:178374471
BUG=b:179795020
TEST=on Jinlon, do the reboot process described in b/179795020#comment3
and see the LTE modem coming back less than 25 second after boot and
attaching to the LTE network.

Cq-Depend: chromium:2703257
Change-Id: I204cb072b1806b463c91cf47594866f70950451b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2704174
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>