overlay-chell: remove is-kernelnext USE flag
This flag is only meant for ${board}-kernelnext overlays which are
involved in an uprev, and not kernelnext profiles in the original
overlay.  The purpose of this flag is to enable the
autoupdate_StatefulCompatibility tauto test, which doesn't make sense
for kernelnext profiles.

BUG=b:181771049
TEST=none

Change-Id: I71ee809e38d12e702ef217d04a0efeab954249f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2733009
Tested-by: Ross Zwisler <zwisler@google.com>
Auto-Submit: Ross Zwisler <zwisler@google.com>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>