overlay-cave: remove is-kernelnext USE flag
This flag is only meant for ${board}-kernelnext overlays which are
involved in an uprev, and not kernelnext profiles in the original
overlay.  The purpose of this flag is to enable the
autoupdate_StatefulCompatibility tauto test, which doesn't make sense
for kernelnext profiles.

BUG=b:181771049
TEST=none

Change-Id: I21263fafb9ff0ce343209515c76d8ec336b84112
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2733010
Tested-by: Ross Zwisler <zwisler@google.com>
Auto-Submit: Ross Zwisler <zwisler@google.com>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>