verify_fsg: Reload FSG atomically
Note: This is essentially a "cherry-pick" of crrev.com/c/2794177, except
we're cherry-picking to a different board overlay, so the path is all
different.

In b/181406065#comment35, it's theorized that some devices failed to
make it through the factory flow because they experienced a power loss
during initial FSG loading. For speed, we do the FSG loading in two
steps: a first step where we copy in 512 byte blocks, and then an
optional second step where we copy any trailing bytes if the file size
isn't a multiple of 512. In comment #28, Jian-Jia found an FSG whose
file size was 980480 instead of the expected 980904, indicating the
second step of the copy never completed, perhaps due to an unexpected
reboot.

To mitigate this, do two things. First, don't restore the file directly
to its final destination. Restore to a temporary file, sync, then
rename the file to its final destination. The sync should act as a
fence ensuring the file is fully written to the disk before the rename
is attempted. The rename itself should be atomic.

Second, Jian-Jia found a nice way to do the dd as a single command and
also get block-sized performance out of it. We'll also check the return
code of dd to explicitly fail if something happened during the copy.

BUG=b:181406065
TEST=rm /var/lib/rmtfs/boot/*, reboot trogdor

Change-Id: I48fb07ddbdcde2ad3c2b53f9d68db3675bd9f825
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2797851
Commit-Queue: Evan Green <evgreen@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Auto-Submit: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>