herobrine: Add sc7280-qc_blobs
The builder is currently hitting an issue with the blobs for sc7280
not being available.  We need a placeholder for the blobs until we get
all the 7280 blobs committed.  Temporarily using the sc7180 blobs for
now.  Will replace when all the blobs are in.

BUG=b:182963902
TEST=emerge-herobrine coreboot

Change-Id: I88c8e305054f045a73493b036f6252486327d3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2841263
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>