Disable WPA3-SAE on Marvell WiFi overlays
I'm going to flip the default to "enabled", because almost all boards
and WiFi chips can support it (and all new ones should support it).
Let's prep by disabling the feature on those where we know it doesn't
work.

It's possible we can make this runtime-detected in the future
(b/184395063), so leave a note about that too.

BUG=b:184395400
TEST=build

Change-Id: I9990d7396a5866d41e028ee7377fa05302e51c15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2840804
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>