strongbad: Keep the USB hub powered during suspend
During system suspend the onboard_usb_hub kernel driver powers its
associated USB hub down, unless a wakeup capable device (typically
HID) is connected. This is generally intended to save power during
suspend when there is no USB device that could wake up the system.
However things are more complicated on devices with a detachable
keyboard/touchpad ('base'). The interface between the board and
the base is USB, both the keyboard and touchpad are wakeup capable.
That doesn't mean though that wakeup is always enabled. Depending
on the 'mode' of the device ('closed', 'laptop', 'tablet, ...)
powerd can disable wakeup for specific USB devices to avoid
undesired wakeup events. In 'closed' mode wakeup from the keyboard
and the touchpad are disabled, as a result the onboard_usb_hub
driver powers the USB hub down when system suspend is entered. This
may seem desirable to save power, but it results in USB resets of
the base, which causes the touchpad to leave its low power mode, in
consequence the power consumption of the base is higher than the
combined consumption of hub + base when the hub remains powered.

Add a udev rule to configure the onboard_usb_hub driver to keep
the USB hub always powered.

BUG=b:183182461
TEST=on coachz:
  cat /sys/bus/platform/devices/usb-hub/always_powered_in_suspend
    => 1

Change-Id: I62ec7f14627cd747c1ebd203a402939f93b62613
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2843940
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>