kukui: modify camera config for kakadu and katsu
The eeprom of the front camera of katsu and kakadu cannot be read
 normally. After testing, it is found that the specification of
Sunny's eeprom is written as A8, but it is actually A0.
The specifications of kingcome's eeprom are written in A4.
In order to facilitate os update, change the i2c address of
eeprom of Sunny and KingCome to A0.

BUG=b:185972351
TEST=cros_workon-kukui start media-libs/cros-camera-hal-mtk
     emerge-kukui media-libs/cros-camera-hal-mtk
     install os and check if data of
     "/run/camera/EEPROM/sub_sensor_eeprom" exists

Cq-Depend: chromium:2887206
Change-Id: Ie2aef38e079061928d29bd272bc587829fc74dd3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2887307
Reviewed-by: Jasmine Chen <lnishan@google.com>
Reviewed-by: Hsu Wei-Cheng <mojahsu@chromium.org>
Tested-by: Zhengqiao Xia <xiazhengqiao@huaqin.corp-partner.google.com>
Commit-Queue: Jasmine Chen <lnishan@google.com>