Enable ondeivce_grammar USE flag on some boards.
Enable on-deivce grammar on
 - atlas
 - chell
 - eve
 - kevin
 - kukui
 - zork

On-device grammar has been enabled on nocturne for more than half a year
and it all went well. Now we want to enable it on more boards. These
boards are used among the team so after enabling them we could do a
teamwide dogfood to verify the feature.

BUG=chromium:1132699
TEST=tested on nocturne, and emerge tests passed for atlas.

Change-Id: If2c97198f216a7604524da515f83c72f3d4efdd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2929609
Reviewed-by: Jing Wang <jiwan@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Tested-by: Jing Wang <jiwan@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>