asurada: enable zephyr-ec build target config
Add a build target for zephyr-ec.  This enables the asurada-zephyr
build to build the correct build targets.

Note: this change is a no-op for the regular asurada build (the
"zephyr" USE flag is required for it to have any effect).

BUG=b:194228475
TEST=merged config for asurada-zephyr:
     https://paste.googleplex.com/5818838126952448

Change-Id: I1b57643ea87a58ffca790b37218300703921d423
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3052694
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>