Delbin/Drobit: Set PreemptHotword for internal mic issue
As mentioned in b:187678153 comment#165, the internal mic failed to
capture while dsp hotwording feature is on for the SOF firmware used in
delbin and drobit.

In order to dogfood on M93, we set PreemptHotword to 1 for Internal Mic
in UCM config of delbin and drobit. Revert this setting after the issue
is fixed.

BUG=b:170281884
BUG=b:187678153
TEST=emerge-volteer chromeos-config-bsp-volteer-private
     chromeos-bsp-volteer chromeos-config

Change-Id: I9cc42b1f45e45be0dcf1f4b4740750a5ebd5cb2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3060119
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>