eve: enable coresched testing
NB: this flag is used only for controlling test behavior; the kernel
configuration is common to all Intel systems.

We should probably migrate this to a hardware dependency that
corresponds to "SMT vulnerable" in
/sys/devices/system/cpu/vulnerabilities/l1tf eventually, but we can at
least extend this usage to the appropriate overlay(s) in the meantime.

BUG=b:195061310
TEST=sched.CoreSchedTag

Change-Id: I875a6a342b75ff0e10a0147116903822a243dc4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3092772
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>