baseboard-gru: Update udev media rules for upstream kernel drivers
The name of the rkvdec video device (used for H.264 decoding) was
changed to simply "rkvdec" in the upstream kernel driver, instead
of based on the compatible string.

Add new entries for the new name in the udev rules file and udev
trigger script.

Also, the upstream drivers now use the media controller API, which has
/dev/media* devices. Add rules for them so that /dev/media-dec* symlinks
are created as well.

Since codec enumeration does not depend on the media devices, triggers
are not added for them.

BUG=b:197576846
TEST=Build 5.10 kernel with CONFIG_STAGING_MEDIA=y and
     CONFIG_VIDEO_ROCKCHIP_VDEC=y; /dev/video-dec0 and /dev/media-dec*
     symlinks should be present

Change-Id: I72787925b40a8dbbc4a63b8b825ef66db86ad455
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3124736
Auto-Submit: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>