Hatch: remove sound-card-init-conf key from models don't use sound_card_init
Remove sound-card-init-conf key from the cros_config for the devices
that don't use sound_card_init. Only NightFury has this key in Hatch
now.

This is to align the logic used in the boxster configurations that if
the sound-card-init-conf key does not exist in the cros_config, it’s
regarded as a device that doesn't not have smart amps.

BUG=b:199343845
TEST=1.`emerge-hatch chromeos-config-bsp-hatch chromeos-config\
        chromeos-bsp-hatch`
     2. Verify that sound-card-init-conf only exists on Nightfury.

Change-Id: Ia4039b484ad67cd2baaac684e7fab9c856506cf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3163998
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>