chipset-apl: Disable Vulkan
Vulkan never had consumers on this chipset, and has no consumers now.
The only potential consumer in the future is the Chrome browser (Skia
and WebGPU), and Vulkan is not directly user-visible there. We can later
re-enable Vulkan for those use cases if we wish. Maintaining Vulkan on
this chipset would be an extra burden due to the old 4.4 kernel.

BUG=none
TEST=none (it's gone, nothing to test)

Change-Id: I6ec5cca3fa8235594b7570b2c8ce003f5ed080d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3179871
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>