Skl-Y: Removing kvm_transition from Skl-Y platforms
In Glados family all Skylake-Y devices are based on kernel v4.19 (in
stable from M91). Because of that we can fully expose Crostini to
users as we do not plan to revert back to old kernel (there is no risk
for data loss due to missing VM support in the reverted kernel).

BUG=b:194936261
TEST=setup_board --board caroline;
     ./build_packages --board=caroline;
     ./build_image --board=caroline
     Install image.
     Check that "Linux beta" is available and Crostini is working
     regardless of setup of enable-experimental-kernel-vm-support in
     chrome://flags

Change-Id: I7d7929fc1fa0400709be3dd45d10a633de579588
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3194338
Reviewed-by: Bob Brandt <brndt@google.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Radoslaw Biernacki <biernacki@google.com>
Tested-by: Radoslaw Biernacki <biernacki@google.com>