Disable Floss bluetooth on all 2gb boards
We are going to enable the Floss bluetooth daemon on all boards except
the 2GB boards which are low on rootfs space. In order to do that, we
will first disable the flags from all 2GB boards and then enable it for
all boards.

BUG=b:201687614
TEST=CQ (should be no-op since this flag wasn't set before)

Change-Id: Idfba2b98bcefd46586f28a98633fa949954014b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3198245
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Anthony Tran <anthonytran@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>