trogdor: Add kernelnext profile
Add a kernelnext profile to match the new way of setting up kernelnext
boards as profiles. This allows us to easily switch to the new kernel by
repointing the 'kernel' symlink in baseboard-trogdor to the kernel
version that we want to use when we're ready to make the switch.

The trogdor-kernelnext board will use the trogdor:kernelnext profile
which is a thin wrapper on top of these profiles that lets us test the
different kernel version.

BUG=b:201110424
TEST=build_packages --board=trogdor-kernelnext
     build_packages --board=trogdor --profile=kernelnext

Cq-Depend: chrome-internal:4183718
Change-Id: Iafe3b755baf0ab1936e850eb20721abd352f0b1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3219874
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Bob Brandt <brndt@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>