overlay-octopus: Fix speaker switch enable sequence
At present, in some octopus board HiFi.conf files, both the
EnableSequence and DisableSequence for the Speaker device turn the
speaker switch off. While this does not cause trouble in kernel v4.14,
it is an incorrect configuration of the Speaker switch, which causes the
built-in speakers not to work on kernel v5.10, for octopus-kernelnext.

Fix the Speaker device EnableSequence in all "HiFi*.conf" files, so that
the Speaker switch is turned ON when the speaker device is in use.

BUG=b:198565161
TEST="Fully tested all audio use cases for octopus and kernelnext on
blooglet/ampton models. Also ran all Tast tests and audio_Aplay test"

Change-Id: I23974603aaf3002d8612f67c32001089d83f1949
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3231138
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Tested-by: Mani Milani <mani@chromium.org>
Commit-Queue: Mani Milani <mani@chromium.org>