Lars: Quickfix: Add kernel module param to prevent HDMI resolution clamping
This patch switch runtime kernel parameter for i915 DRM
module which helps to overcome regression introduced in
mainline kernel 4.19:
commit d61992565bd3 ("drm/i915: Determine DP++ type 1 DVI
adaptor presence based on VBT")

Mentioned commit changes the behavior for DP dual mode
detection and assume that HDMI port has a DVI type 1 adapter
connected, even though this is not actually the case. This
results in that the connected HDMI output is limited in
frequency which limits the available refresh rates and
resolution and therefore worsening user experience. This is
considered regression for Chromebooks with built in HDMI
output as user is unable to use full capabilities of
widescreen monitors.

This fix should be treated as a HACK till we land proper
solution in upstream, but is needed as short term solution
for P0 BUG for Lars on M94.

BUG=b:203378940
TEST=Build and flash Lars recovery image.
     Connect HDMI widescreen monitor.
     ChromeOs should use full set of monitor resolutions.

Cq-Depend: chrome-internal:4203117
Change-Id: I4a8420d3c10b97ba4668264f866bdb8f15aeef6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3236995
Reviewed-by: Bob Brandt <brndt@google.com>
Reviewed-by: Marcin Wojtas <mwojtas@google.com>
Tested-by: Radoslaw Biernacki <biernacki@google.com>
Auto-Submit: Radoslaw Biernacki <biernacki@google.com>
Commit-Queue: Marcin Wojtas <mwojtas@google.com>