ml: disable text suggestions on 2gb boards
This CL ensures the text suggester model housed within the ml service is
disabled for all 2gb boards as they are not supported.

This is a prerequisite to an upcoming CL where we enable the text
suggester model in all boards where the ml service is enabled, with 2gb
boards being the exception.

BUG=chromium:1146266
TEST=No affected tests because the flag was disabled by default.

Change-Id: Ife834e935b0c80db0ea704b6e9539a3fb84d1dfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3239707
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Jing Wang <jiwan@chromium.org>