project-manatee: Add VFIO hack to allow intel-lpss driver
The kernel checks VFIO groups for "viability". Intuitively, this
determines whether it is safe to operate hardware devices from
userspace via VFIO. One condition is for all devices in the VFIO group
to not have a (non-trivial) driver loaded. This is necessary to
guarantee memory isolation, since all devices in the IOMMU group can
see all mapped memory, and buffers mapped by a bound kernel driver
would thus become indirectly accessible to userspace.

intel-lpss is the driver that handles PCI serial port hardware on
Intel systems. This change adds it to the allowlist. This is safe
because the driver does not set up any IOMMU mappings, in fact it only
operates on the MMIO registers exposed via PCI BAR 0. Note that the
serial driver technically has the ability to make use of DMA, but
fortunately that is not used in our case (hence calling this change a
hack that is not appropriate to propose upstream).

BUG=b:167947780
TEST=VFIO works even in groups where the intel-lpss driver is still bound.

Change-Id: I72a2c733057fd059384342befb7979513f4a748d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3247174
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>