amd64-generic: Add unibuild config
Add a minimal model.yaml to amd64-generic overlay.  This enables
chromeos-config for the overlay, even though the board isn't intended
to be used in a typical unibuild fashion.

The purpose of this is to eventually enable unibuild globally,
including in generic overlays and VMs.  See go/unibuild-migration for
greater context.

BUG=b:187699715
TEST=build image for both amd64-generic and novato
     boot images using cros_vm and verify "cros_config / name" works
     as intended

Cq-Depend: chrome-internal:4264557,chrome-internal:4264558
Change-Id: I50890a9f87003dddfad84075ea80a52713f9ae69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3284305
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>