kevin: Swap DP/HDMI and Wake-on-Voice audio config for new kernel
Copy the original audio configs to audio-config-4.4, and make one tweak
for kernel 5.10+ in audio-config/.

The ordering of the sub-devices with rk3399-gru-sound is dependent on
the order of the codecs listed in the DTS file. When upstreamed, this
ordering changed to list the Wake-on-Voice codec before the DP/HDMI
audio codec.

Drop "rk3399-gru-sound DP Jack" while we're at it, since it didn't make
it upstream yet: b/207015460.

BUG=b:205667230, b:207015460
TEST=audio_AudioBasicHotwording, on kevin and kevin-kernelnext

Change-Id: I04eb0832c95ed371617012919af57e5118879438
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3291897
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>