starnix: Create new profile for Fuchsia Starnix
The Fuchsia Starnix profile is meant to only contain Chromium OS' rootfs.

It starts out with only a few real binaries and no Chrome. This is to
help focus on getting a handful of Chromium OS' dependencies running
successfully.

Eventually, the plan is to encompass more of the real Chromium OS image.

BUG=chromium:209505710
TEST=Built image

Change-Id: I71eb14f92c13600a365b9c5506e53b88bceec37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3296703
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>