lakitu: drop board
Development on this project has been forked to sep source trees and
this is no longer being updated.  Drop it from the tree to avoid
confusion and because it's using APIs that we want to delete and/or
change.  No sense in migrating this usage if no one is using it.

BUG=b:207034476
TEST=CQ passes

Change-Id: Idce82ed12afd93505c2483422dda1570a6c20133
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3336284
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Owners-Override: Mike Frysinger <vapier@chromium.org>