senor: Remove EC use flags
We enabled the EC use flags in baseboard-herobrine but senor doesn't have
an EC board. The senor build fails now because we try to build the EC code
but we can't. Remove the EC use flags from senor so that build packages keep
working.

BUG=b:175846610, b:212665845
TEST=build_packages --board=senor

Change-Id: Ib91883865dd42d5f23be28cfcbeefb6d2819d19b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3387811
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>