baseboard_brya: Add "intel_iommu=on" to commandline again
We had removed this from commandline due to an IOMMU in ES silicon.
This is fixed in QS silicon and now we (are begining to) have a mix
of ES and QS silicon. Lets re-enable this flag in the kernel. Now
the IOMMU shall be enabled in the kernel as long as it is enabled in
BIOS.

Since currently BIOS disables IOMMU for ES silicon, this should have no
regression on ES silicon. BIOS already enables IOMMU for QS silicon, so
we shall have it enabled there

BUG=b:197177091
TEST=Check that IOMMU domains are not enabled on ES silicon, and enabled
on QS silicon.

Change-Id: Ibf09a0f62b2114938475111767cda49f84ca15d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/3426163
Reviewed-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>